requirements.txt
transformers 
streamlit 
matplotlib 
seaborn 
wordcloud
pandas
numpy
emoji
tensorflow
torch
torchvision
torchaudio
PyPDF2
python-docx
logging
pathlib
tempfile
re
collections
nltk
tf.keras